bug 1787931: move super search models to webapp
The webapp uses a couple of models that were defined in
socorro/external/es/super_search_fields.py, but aren't used in socorro/
and have a lot of webapp-centric things in them. I moved them to the
webapp into a libsupersearch module that'll act as an intermediary layer
between the webapp and the rest of socorro.
